+++ Example/iOSDFULibrary/View Controllers/DFUViewController.swift
@@ -191,6 +191,15 @@ class DFUViewController: UIViewController, CBCentralManagerDelegate, DFUServiceD
         dfuInitiator.delegate = self
         dfuInitiator.progressDelegate = self
         dfuInitiator.logger = self
+        dfuInitiator.dataObjectPreparationDelay = 0.4 // sec
+        
+        // Uncomment if you don't want resume feature in Secure DFU.
+        // See: https://github.com/NordicSemiconductor/IOS-Pods-DFU-Library/pull/264
+        // dfuInitiator.disableResume = true
+        
+        // Uncomment only if Legacy DFU is us using address+1 in bootloader mode.
+        // See: https://github.com/NordicSemiconductor/Android-DFU-Library/issues/262#issuecomment-665493850
+        // dfuInitiator.forceScanningForNewAddressInLegacyDfu = true
 
         // Here would be a good chance to change the UUIDs to your custom UUIDs